move _extract_privileges to Grantable, add role drop to test_database's revoke test, fill out grant/revoke tests for schema
