+++ tests/it/test_schema.py
@@ -2,6 +2,7 @@ import pytest
 from sqlalchemy import Engine
 
 from postgres_declare.base import Base
+from postgres_declare.data_structures.grant_to import GrantTo
 from postgres_declare.data_structures.privileges import Privilege
 from postgres_declare.entities.database import Database
 from postgres_declare.entities.entity import Entity
@@ -27,17 +28,23 @@ def test_create(simple_schema: Schema) -> None:
 
 @pytest.mark.order(after="test_create")
 def test_grant_does_not_exist(simple_schema: Schema, grant_role: Role, schema_privileges: set[Privilege]) -> None:
-    pass
+    grant_role._safe_create()
+    assert not simple_schema._grants_exist(grantee=grant_role, privileges=schema_privileges)
 
 
 @pytest.mark.order(after="test_grant_does_not_exist")
 def test_grant(simple_schema: Schema, grant_role: Role, schema_privileges: set[Privilege]) -> None:
-    pass
+    simple_schema.grant(grants=[GrantTo(privileges=list(schema_privileges), to=[grant_role])])
+    grant_role._safe_grant()
+    assert simple_schema._grants_exist(grantee=grant_role, privileges=schema_privileges)
 
 
 @pytest.mark.order(after="test_grant")
 def test_revoke(simple_schema: Schema, grant_role: Role, schema_privileges: set[Privilege]) -> None:
-    pass
+    grant_role._safe_revoke()
+    assert not simple_schema._grants_exist(grantee=grant_role, privileges=schema_privileges)
+    # clean up role
+    grant_role._safe_drop()
 
 
 @pytest.mark.order(after="test_create")
